apparently everything wants to break now, goddamnit

+++ src/main/java/io/github/Gabriel/NMLSkills/energySystem/EnergyListener.java
@@ -18,7 +18,7 @@ public class EnergyListener implements Listener {
     public void onFoodLevelChange(FoodLevelChangeEvent event) {
         if (event.getEntity() instanceof Player player) {
             int foodLevelChange = event.getFoodLevel() - player.getFoodLevel();
-            double energyChange = foodLevelChange * (nmlAttributes.getProfileManager().getPlayerProfile(player.getUniqueId()).getAttributes().getMaxEnergy() / 20);
+            int energyChange = foodLevelChange * (nmlAttributes.getProfileManager().getPlayerProfile(player.getUniqueId()).getAttributes().getMaxEnergy() / 20);
 
             nmlAttributes.getProfileManager().getPlayerProfile(player.getUniqueId()).getAttributes().setCurrentEnergy(nmlAttributes.getProfileManager().getPlayerProfile(player.getUniqueId()).getAttributes().getCurrentEnergy() + energyChange);
         }
@@ -27,7 +27,7 @@ public class EnergyListener implements Listener {
     @EventHandler
     public void onPlayerMove(PlayerMoveEvent event) {
         Player player = event.getPlayer();
-        double maxEnergy = nmlAttributes.getProfileManager().getPlayerProfile(player.getUniqueId()).getAttributes().getMaxEnergy();
+        int maxEnergy = nmlAttributes.getProfileManager().getPlayerProfile(player.getUniqueId()).getAttributes().getMaxEnergy();
 
         if (player.isSprinting()) {
             nmlAttributes.getEnergyManager().useEnergy(player, maxEnergy / 200);
